refactor CLI integration tests
